Delay report of external displays until the primary gets enabled
On Oreo, SurfaceFlinger does not enable the external display
if the hotplug event is sent directly after registering the callback.
So if the external screen is already connected while booting, it is
never enabled until it is hotplugged once. It works fine on Pie.

Workaround this problem by delaying the report of the external
display(s) until the primary display gets enabled.